added test, and modified data
