+++ _data/test.yml
@@ -88,6 +88,3 @@ id_journalpaper_tochi2005_fogarty_james:
   title: Predicting Human Interruptibility with Sensors
   volume: 12
   year: 2005
-
-
-
